refactor: map code splitting #94
